add Exceptions, log4j, change int to BigInteger, do cosmetic fix.
Created new test methods
Implements cannonDelete method

+++ src/main/java/com/nctc2017/dao/ExecutorDao.java
@@ -1,21 +1,20 @@
 package com.nctc2017.dao;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
 public interface ExecutorDao {
 
-	boolean ifThingBelongToPlayer(int id, int idPerson);
+    boolean ifThingBelongToPlayer(int id, int idPerson);
 
-	boolean calculateDamage(List<List<Integer>> ammoCannon, int idMyShip, int idEnemyShip);
+    boolean calculateDamage(List<List<Integer>> ammoCannon, int idMyShip, int idEnemyShip);
 
-	int boarding(int idMyShip, int idEnemyShip);
+    int boarding(int idMyShip, int idEnemyShip);
 
-	void moveCargoTo(int cargoId, int destinationId, int quantity);
+    void moveCargoTo(int cargoId, int destinationId, int quantity);
 
-	String moveCargoToWinner(int shipWinnerId, int shipLosserId);
-	
-	BigDecimal createCannon(int templateId);
+    String moveCargoToWinner(int shipWinnerId, int shipLosserId);
+
+    BigInteger createCannon(int templateId);
 
 }
